[SB-03]Añadido diseño de elementos del recyclerview y implementado método para obtener los logros de la BD. La consulta devuelve un error, por lo que debo investigar a que se debe. El SB-04 se dedicará a solucionar este error con carácter urgente.

+++ app/src/main/java/icarus/silver/scoreboards/models/Juego.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 
 public class Juego implements Serializable {
 
-    private int idJuego;
+    private long idJuego;
     private String nombre;
     private float precio;
     private int rebaja;
@@ -29,11 +29,11 @@ public class Juego implements Serializable {
     public Juego() {
     }
 
-    public int getIdJuego() {
+    public long getIdJuego() {
         return idJuego;
     }
 
-    public void setIdJuego(int idJuego) {
+    public void setIdJuego(long idJuego) {
         this.idJuego = idJuego;
     }
